new product get query, category ancestor exception

+++ tgext/ecommerce/lib/category.py
@@ -2,7 +2,7 @@
 from __future__ import unicode_literals
 from bson import ObjectId
 import tg
-from tgext.ecommerce.lib.exceptions import CategoryAssignedToProductException
+from tgext.ecommerce.lib.exceptions import CategoryAssignedToProductException, CategoryAcestorExistingException
 from tgext.ecommerce.lib.utils import slugify, internationalise as i_, NoDefault, slugify_category
 from tgext.ecommerce.model import models
 
@@ -50,16 +50,18 @@ class CategoryManager(object):
 
         for cat in models.Category.query.find({'ancestors._id': ObjectId(_id)}):
             parent = models.Category.query.find({'_id': cat.parent}).first()
-            ancestors = [ancestor for ancestor in parent.ancestors]
-            ancestors.append(dict(_id=parent._id, name=parent.name, slug=parent.slug))
-            models.Category.query.update({'_id':cat._id}, {'$set': {'ancestors': ancestors}})
+            if parent:
+                ancestors = [ancestor for ancestor in parent.ancestors]
+                ancestors.append(dict(_id=parent._id, name=parent.name, slug=parent.slug))
+                models.Category.query.update({'_id':cat._id}, {'$set': {'ancestors': ancestors}})
 
 
     @classmethod
     def delete(cls, _id): #delete_category
         if models.Product.query.find({'category_id': ObjectId(_id), 'active': True}).first():
             raise CategoryAssignedToProductException('The Category is assigned to an active Product')
-
+        if models.Category.query.find({'ancestors._id': ObjectId(_id)}).first():
+            raise CategoryAcestorExistingException
         models.Category.query.get(_id=ObjectId(_id)).delete()
         models.Product.query.update({'category_id': ObjectId(_id), 'active': False},
                                     {'$set': {'category_id': None}})
